Add PKCS12_check_version API
Fixes #22030

There was no check for the version = 3 when loading
a PKCS12 file. Rather than adding multiple _ex() functions
that support checking the version number a separate API has been
added to do the check which can be called after one of the d2i_PKCS12
functions.
The check has been added to the PKCS12 commandline application. It is
checked by default and can be disabled by using the option -noversioncheck.